+++ src/features/foldingProvider.ts
@@ -6,7 +6,7 @@
 import { Token } from '../../libs/textile-js/textile';
 import * as vscode from 'vscode';
 import { TextileEngine } from '../textileEngine';
-import { TableOfContentsProvider } from '../tableOfContentsProvider';
+import { TableOfContents } from '../tableOfContentsProvider';
 
 const rangeLimit = 5000;
 
@@ -50,7 +50,7 @@ export default class TextileFoldingProvider implements vscode.FoldingRangeProvid
 			}]
 		]);
 
-		const nestingStack: { line: number, isStart: boolean }[] = [];
+		const nestingStack: { line: number; isStart: boolean }[] = [];
 		return regionMarkers
 			.map(marker => {
 				if (marker.isStart) {
@@ -67,9 +67,8 @@ export default class TextileFoldingProvider implements vscode.FoldingRangeProvid
 	// --- End : modified for textile
 
 	private async getHeaderFoldingRanges(document: vscode.TextDocument) {
-		const tocProvider = new TableOfContentsProvider(this.engine, document);
-		const toc = await tocProvider.getToc();
-		return toc.map(entry => {
+		const toc = await TableOfContents.create(this.engine, document);
+		return toc.entries.map(entry => {
 			let endLine = entry.location.range.end.line;
 			if (document.lineAt(endLine).isEmptyOrWhitespace && endLine >= entry.line + 1) {
 				endLine = endLine - 1;